requirements.txt
alembic==1.14.0
anyio==4.7.0
APScheduler==3.11.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
decorator==4.4.2
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
facebook-sdk==3.1.0
Flask==3.1.0
Flask-Migrate==4.0.7
Flask-MySQLdb==2.0.0
flask-paginate==2024.4.12
Flask-Seeder==1.2.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
moviepy==1.0.3
mysql-connector-python==9.1.0
mysqlclient==2.2.6
numpy==2.2.1
openpyxl==3.1.5
pandas==2.2.3
pillow==10.4.0
proglog==0.1.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.9
pytz==2024.2
requests==2.32.3
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1
yt-dlp==2024.12.23
